requirements.txt
asgiref==3.2.5
Babel==2.8.0
Django==3.0.4
django-crispy-forms==1.9.0
django-filter==2.2.0
django-phonenumber-field==4.0.0
djangorestframework==3.11.0
phonenumbers==8.11.5
Pillow==7.0.0
pytz==2019.3
sorl-thumbnail==12.6.3
sqlparse==0.3.1
